requirements.txt
streamlit
pypdf
pymupdf
Pillow
